Working menu on mobile and desktop
