+++ components/organisms/Layout.stories.js
@@ -24,6 +24,7 @@ const Template = (args) => <Layout {...args}>{sampleText()}</Layout>;
 
 export const NoBanner = Template.bind({});
 export const WithBanner = Template.bind({});
+export const WithMenu = Template.bind({});
 
 NoBanner.args = {
   langUrl: "someUrl",
@@ -33,6 +34,12 @@ NoBanner.args = {
 WithBanner.args = {
   bannerTitle: "Banner Title",
   bannerText: "Banner Text",
+  langUrl: "someUrl",
+  locale: "someLocale",
+};
+
+WithMenu.args = {
+  menu: true,
   langUrl: "someUrl",
   locale: "someLocale",
 };
