+++ components/organisms/Layout.test.js
@@ -1,7 +1,7 @@
 import { render, screen } from "@testing-library/react";
 import "@testing-library/jest-dom/extend-expect";
 import { axe, toHaveNoViolations } from "jest-axe";
-import { NoBanner, WithBanner } from "./Layout.stories";
+import { NoBanner, WithBanner, WithMenu } from "./Layout.stories";
 
 expect.extend(toHaveNoViolations);
 
@@ -14,6 +14,11 @@ describe("Layout", () => {
   it("renders with the banner", () => {
     render(<WithBanner {...WithBanner.args} />);
     screen.getByTitle("Home banner");
+  });
+
+  it("renders with the menu", () => {
+    render(<WithBanner {...WithMenu.args} />);
+    screen.getByTitle("Menu");
   });
 
   it("has no a11y violations", async () => {
